feat:🪝standarizes useQueryTab into hook.
