+++ pages/settings.tsx
@@ -5,17 +5,15 @@ import {
   TabList,
   TabPanels,
   Tabs,
-  useColorMode,
+  useColorMode
 } from '@chakra-ui/react';
-import { useSearchParams } from 'next/navigation';
-import { useRouter } from 'next/router';
-import { useCallback, useEffect, useState } from 'react';
 
 import ConnectionsTab from '@/components/editor/settings/ConnectionsTab';
 import DangerZoneTab from '@/components/editor/settings/DangerZoneTab';
 import OverviewTab from '@/components/editor/settings/OverviewTab';
 import ProfileTab from '@/components/editor/settings/ProfileTab';
 import SessionsTab from '@/components/editor/settings/SessionsTab';
+import useQueryTab from '@/hooks/useQueryTab';
 
 const TabIndexes: { [key in string]: number } = {
   overview: 0,
@@ -32,47 +30,8 @@ export default function Settings() {
   const { colorMode } = useColorMode();
   const bg = colorMode === 'light' ? 'blackAlpha.300' : 'blackAlpha.700';
 
-  const searchParams = useSearchParams();
-  const { query, push: navigateTo } = useRouter();
-  const [tabIndex, setTabIndex] = useState(0);
+  const { tabIndex, setTabIndex } = useQueryTab(TabIndexes);
 
-  // Control query.
-  const createQueryString = useCallback(
-    (name: string, value: string | null) => {
-      const params = new URLSearchParams(searchParams.toString());
-
-      if (value) {
-        params.set(name, value);
-      } else {
-        params.delete(name);
-      }
-
-      return params.toString();
-    },
-    [searchParams],
-  );
-
-  useEffect(() => {
-    const tab = searchParams.get('tab');
-    if (tab) {
-      const newIndex = TabIndexes[tab] || 0;
-      setTabIndex(newIndex);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    for (const [key, value] of Object.entries(TabIndexes)) {
-      if (value === tabIndex) {
-        const newQuery = createQueryString(
-          'tab',
-          key === 'overview' ? null : key,
-        );
-        navigateTo({ search: newQuery });
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [tabIndex]);
 
   return (
     <Flex flexDir={'column'} gap={'10px'} width={'100%'}>
